requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.6
Werkzeug==2.3.7
psycopg2-binary==2.9.7
python-dotenv==1.0.0
bcrypt==4.0.1
gunicorn==21.2.0
email-validator==2.0.0
Pillow==10.0.0
redis==4.6.0
celery==5.3.1
reportlab==4.0.4
XlsxWriter==3.1.9
matplotlib==3.7.2
seaborn==0.12.2
pandas==2.1.1
python-socketio==5.8.0
pytest==7.4.2
pytest-flask==1.2.0
